fix: Pass in char const* to sql::bytes constructor instead of char *.
sql::bytes have different constructors when we pass in char * and char
const* together with the length. When we pass in char *, the length it
stored internally is -length we passed in. To get the correct behavior.
i.e. having a valid byte array with positive range, we must pass in char
const*.