fix: Lock task to v3.39.2 to work around behaviour change in v3.40.0. (#22)
The GH linting workflow was failing since it was using the latest version (v3.40.0) of task which changes the behaviour of how undefined variables are treated in ref statements, causing yscope-dev-utils' validate-checksum task to fail.

Locking task's version to v3.39.2 to work around the issue. Once yscope-dev-utils' tasks are updated, we can remove this constraint.